Send 403 if contact does not belong to user

+++ delete.php
@@ -20,6 +20,15 @@ if ($statement->rowCount() == 0) {
   return;
 }
 
+// Not allow to delete other users contacts
+
+$contact = $statement->fetch(PDO::FETCH_ASSOC);
+
+if ($contact["user_id"] !== $_SESSION["user"]["id"]) {
+  http_response_code(403);
+  echo("HTTP 403 UNAUTHORIZED");
+  return;
+}
 
 $conn->prepare("DELETE FROM contacts WHERE id = :id")->execute([":id" => $id]);
